+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/components/Apis/Details/Documents/Download.jsx
@@ -23,59 +23,18 @@ import Button from '@material-ui/core/Button';
 import Icon from '@material-ui/core/Icon';
 import Alert from 'AppComponents/Shared/Alert';
 import Api from 'AppData/api';
+import Utils from 'AppData/Utils';
 
 function Download(props) {
     const { intl } = props;
 
     const { docId, apiId } = props;
-    /**
-     * Download the document related file
-     * @param {any} response Response of download file
-     */
-    const downloadFile = (response, doc) => {
-        let fileName = '';
-        const contentDisposition = response.headers['content-disposition'];
-
-        if (contentDisposition && contentDisposition.indexOf('attachment') !== -1) {
-            const fileNameReg = /filename[^;=\n]*=((['"]).*?\2|[^;\n]*)/;
-            const matches = fileNameReg.exec(contentDisposition);
-            if (matches != null && matches[1]) fileName = matches[1].replace(/['"]/g, '');
-        }
-        const contentType = response.headers['content-type'];
-        const blob = new Blob([response.data], {
-            type: contentType,
-        });
-        if (typeof window.navigator.msSaveBlob !== 'undefined') {
-            window.navigator.msSaveBlob(blob, fileName);
-        } else {
-            const URL = window.URL || window.webkitURL;
-            const downloadUrl = URL.createObjectURL(blob);
-
-            if (fileName) {
-                const aTag = document.createElement('a');
-                if (typeof aTag.download === 'undefined') {
-                    window.location = downloadUrl;
-                } else {
-                    aTag.href = downloadUrl;
-                    aTag.download = fileName;
-                    document.body.appendChild(aTag);
-                    aTag.click();
-                }
-            } else {
-                window.location = downloadUrl;
-            }
-
-            setTimeout(() => {
-                URL.revokeObjectURL(downloadUrl);
-            }, 100);
-        }
-    };
     const handleDownload = () => {
         const api = new Api();
         const promised_get_content = api.getFileForDocument(apiId, docId);
         promised_get_content
-            .then((done) => {
-                downloadFile(done, document);
+            .then((response) => {
+                Utils.forceDownload(response);
             })
             .catch((error) => {
                 if (process.env.NODE_ENV !== 'production') {
