+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/components/Apis/Details/Documents/View.jsx
@@ -35,6 +35,8 @@ import Alert from 'AppComponents/Shared/Alert';
 import API from 'AppData/api';
 import APIProduct from 'AppData/APIProduct';
 import APIContext from 'AppComponents/Apis/Details/components/ApiContext';
+import Utils from 'AppData/Utils';
+
 const ReactMarkdown = lazy(() => import('react-markdown' /* webpackChunkName: "ViewReactMD" */));
 
 const styles = theme => ({
@@ -134,53 +136,12 @@ function View(props) {
                 }
             });
     };
-    /**
-     * Download the document related file
-     * @param {any} response Response of download file
-     */
-    const downloadFile = (response, doc) => {
-        let fileName = '';
-        const contentDisposition = response.headers['content-disposition'];
 
-        if (contentDisposition && contentDisposition.indexOf('attachment') !== -1) {
-            const fileNameReg = /filename[^;=\n]*=((['"]).*?\2|[^;\n]*)/;
-            const matches = fileNameReg.exec(contentDisposition);
-            if (matches != null && matches[1]) fileName = matches[1].replace(/['"]/g, '');
-        }
-        const contentType = response.headers['content-type'];
-        const blob = new Blob([response.data], {
-            type: contentType,
-        });
-        if (typeof window.navigator.msSaveBlob !== 'undefined') {
-            window.navigator.msSaveBlob(blob, fileName);
-        } else {
-            const URL = window.URL || window.webkitURL;
-            const downloadUrl = URL.createObjectURL(blob);
-
-            if (fileName) {
-                const aTag = document.createElement('a');
-                if (typeof aTag.download === 'undefined') {
-                    window.location = downloadUrl;
-                } else {
-                    aTag.href = downloadUrl;
-                    aTag.download = fileName;
-                    document.body.appendChild(aTag);
-                    aTag.click();
-                }
-            } else {
-                window.location = downloadUrl;
-            }
-
-            setTimeout(() => {
-                URL.revokeObjectURL(downloadUrl);
-            }, 100);
-        }
-    };
     const handleDownload = () => {
         const promised_get_content = restAPI.getFileForDocument(api.id, documentId);
         promised_get_content
-            .then(done => {
-                downloadFile(done, document);
+            .then(response => {
+                Utils.forceDownload(response);
             })
             .catch(error => {
                 if (process.env.NODE_ENV !== 'production') {
