+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/data/Utils.js
@@ -260,6 +260,53 @@ class Utils {
             return 'rgb(' + r + ', ' + g + ', ' + b + ')';
         }
     }
+
+
+    /**
+     * Force file download in browser
+     *
+     * @static
+     * @param {*} response
+     * @memberof Utils
+     */
+    static forceDownload(response) {
+        let fileName = '';
+        const contentDisposition = response.headers['content-disposition'];
+
+        if (contentDisposition && contentDisposition.indexOf('attachment') !== -1) {
+            const fileNameReg = /filename[^;=\n]*=((['"]).*?\2|[^;\n]*)/;
+            const matches = fileNameReg.exec(contentDisposition);
+            if (matches != null && matches[1]) fileName = matches[1].replace(/['"]/g, '');
+        }
+        const contentType = response.headers['content-type'];
+        const blob = new Blob([response.data], {
+            type: contentType,
+        });
+        if (typeof window.navigator.msSaveBlob !== 'undefined') {
+            window.navigator.msSaveBlob(blob, fileName);
+        } else {
+            const URL = window.URL || window.webkitURL;
+            const downloadUrl = URL.createObjectURL(blob);
+
+            if (fileName) {
+                const aTag = document.createElement('a');
+                if (typeof aTag.download === 'undefined') {
+                    window.location = downloadUrl;
+                } else {
+                    aTag.href = downloadUrl;
+                    aTag.download = fileName;
+                    document.body.appendChild(aTag);
+                    aTag.click();
+                }
+            } else {
+                window.location = downloadUrl;
+            }
+
+            setTimeout(() => {
+                URL.revokeObjectURL(downloadUrl);
+            }, 100);
+        }
+    }
 }
 
 Utils.CONST = {
